vanilla/travel-video-analyzer/requirements.txt
# Core
anthropic>=0.40.0
python-dotenv>=1.0.0
opencv-python>=4.8.0

# OpenTelemetry / OpenInference
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp-proto-http>=1.20.0
openinference-instrumentation>=0.1.0
openinference-instrumentation-anthropic>=0.1.0
openinference-semantic-conventions>=0.1.0
